Move norm files to builder directory.
