+++ src/ngraph/builder/norm.cpp
@@ -28,7 +28,7 @@
 
 namespace ngraph
 {
-    namespace op
+    namespace builder
     {
         namespace detail
         {
@@ -36,7 +36,7 @@ namespace ngraph
                                                        const Shape& node_shape,
                                                        float bias)
             {
-                return Constant::create(
+                return op::Constant::create(
                     element_type, node_shape, std::vector<float>(shape_size(node_shape), bias));
             }
 
@@ -45,49 +45,49 @@ namespace ngraph
                                           const AxisSet& reduction_axes,
                                           float bias)
             {
-                std::shared_ptr<Node> abs_values{std::make_shared<Abs>(node)};
-                std::shared_ptr<Node> p_node = Constant::create(
+                std::shared_ptr<Node> abs_values{std::make_shared<op::Abs>(node)};
+                std::shared_ptr<Node> p_node = op::Constant::create(
                     node->get_element_type(),
                     node->get_shape(),
                     std::vector<float>(shape_size(node->get_shape()), static_cast<float>(p_norm)));
 
-                std::shared_ptr<Node> values{std::make_shared<Power>(abs_values, p_node)};
+                std::shared_ptr<Node> values{std::make_shared<op::Power>(abs_values, p_node)};
 
-                values = std::make_shared<Sum>(values, reduction_axes);
+                values = std::make_shared<op::Sum>(values, reduction_axes);
 
                 std::shared_ptr<Node> bias_node{
                     detail::get_bias_node(values->get_element_type(), values->get_shape(), bias)};
                 values = values + bias_node;
 
-                std::shared_ptr<Node> inv_p_node = Constant::create(
+                std::shared_ptr<Node> inv_p_node = op::Constant::create(
                     values->get_element_type(),
                     values->get_shape(),
                     std::vector<float>(shape_size(values->get_shape()), 1.f / p_norm));
 
-                return {std::make_shared<Power>(values, inv_p_node)};
+                return {std::make_shared<op::Power>(values, inv_p_node)};
             }
         }
 
         std::shared_ptr<Node> l0_norm(const std::shared_ptr<Node>& node,
                                       const AxisSet& reduction_axes)
         {
-            std::shared_ptr<Node> abs_values{std::make_shared<Abs>(node)};
+            std::shared_ptr<Node> abs_values{std::make_shared<op::Abs>(node)};
             std::shared_ptr<Node> zero_node{
-                Constant::create(node->get_element_type(),
+                op::Constant::create(node->get_element_type(),
                                  node->get_shape(),
                                  std::vector<float>(shape_size(node->get_shape()), 0.f))};
 
-            std::shared_ptr<Node> non_zero_values = std::make_shared<Convert>(
-                std::make_shared<NotEqual>(abs_values, zero_node), abs_values->get_element_type());
+            std::shared_ptr<Node> non_zero_values = std::make_shared<op::Convert>(
+                std::make_shared<op::NotEqual>(abs_values, zero_node), abs_values->get_element_type());
 
-            return std::make_shared<Sum>(non_zero_values, reduction_axes);
+            return std::make_shared<op::Sum>(non_zero_values, reduction_axes);
         }
 
         std::shared_ptr<Node>
             l1_norm(const std::shared_ptr<Node>& node, const AxisSet& reduction_axes, float bias)
         {
             std::shared_ptr<Node> values{
-                std::make_shared<Sum>(std::make_shared<Abs>(node), reduction_axes)};
+                std::make_shared<op::Sum>(std::make_shared<op::Abs>(node), reduction_axes)};
 
             std::shared_ptr<Node> bias_node{
                 detail::get_bias_node(values->get_element_type(), values->get_shape(), bias)};
@@ -98,14 +98,13 @@ namespace ngraph
         std::shared_ptr<Node>
             l2_norm(const std::shared_ptr<Node>& node, const AxisSet& reduction_axes, float bias)
         {
-            std::shared_ptr<Node> abs_values{std::make_shared<Abs>(node)};
-            std::shared_ptr<Node> values{
-                std::make_shared<Sum>(abs_values * abs_values, reduction_axes)};
+            // std::shared_ptr<Node> abs_values{std::make_shared<op::Abs>(node)};
+            std::shared_ptr<Node> values{std::make_shared<op::Sum>(node * node, reduction_axes)};
 
             std::shared_ptr<Node> bias_node{
                 detail::get_bias_node(values->get_element_type(), values->get_shape(), bias)};
 
-            return {std::make_shared<Sqrt>(values + bias_node)};
+            return {std::make_shared<op::Sqrt>(values + bias_node)};
         }
 
         std::shared_ptr<Node> lp_norm(const std::shared_ptr<Node>& node,
@@ -135,6 +134,6 @@ namespace ngraph
             }
         }
 
-    } // namespace op
+    } // namespace builder
 
 } // namespace ngraph
